// AppCovid-19/src/models/sitio-inegi.model.ts
export class SitioInegiModel {
    Id: string;
    Nombre: string;
    Razon_social: string;
    Clase_actividad: string;
    Estrato: string;
    Tipo_vialidad: string;
    Calle: string;
    Num_Exterior: string;
    Num_Interior: string;
    Colonia: string;
    CP: string;
    Ubicacion: string;
    Telefono: string;
    Correo_e: string;
    Sitio_internet: string;
    Tipo: string;
    Longitud: string;
    Latitud: string;
    CentroComercial: string;
    TipoCentroComercial: string;
    NumLocal: string;
}