// AppCovid-19/src/models/modelTickets.ts
export class modelTickets {

    idTicket: number;
    idInstancia: number;
    idUsuario: number;
    idCaso: number;
    idCanal: number;
    idSubServicio: number;
    descripcion: string;
    ubicacion: string;
    domicilio: string;
    entidad: string;
    municipio: string;
    colonia: string;
    fechaCreacion: string;
    FechaAtencion: Date;
    folio: number;
    dependenciaDescripcion: string;
    servicioDescripcion: string;
    subServicioDescripcion: string;
    nombreCaso: string;
    comentario: string;
    comentarios: string;
    adjuntos: string;
    idEstatus: number;
    descripcionCanales: string;
    nombreSubServicio: string;
    idEstado: string;
    idMunicipio: string;
    codigoPostal: string;
}