// AppCovid-19/src/models/paciente.model.ts
export class PacienteModel {
    idTicket: number;
    idInstancia: number;
    idCaso: number;
    idCanal: number;
    ubicacionTicket: string;
    folio: string;
    idUsuario: number;
    idTipoUsuario: number;
    nombreUsuario: string;
    telefonoUsuario: string;
    emailUsuario: string;
    calleUsuario: string;
    numeroUsuario: string;
    municipioUsuario: string;
    codigoPostalUsuario: string;
    fechaNacimientoUsuario: string;
    fechaAsgUsuarioTicket: string;
    nombreCaso: string;
    descripcionCaso: string;
    iconCaso: string;
    descripcionCanal: string;
    tipo: string;
    descripcionHospital: string;
    nombreHospital: string;
    ubicacionHospital: string;
    idSubServicio: number;
}